requirements.txt
h5py
hfda
hydra-core
ipykernel
networkx
numpy
pandas
qiskit==2.2.0
qiskit-aer==0.17.0
qiskit-algorithms==0.4.0
qiskit-ibm-runtime==0.44.0
qiskit-ibm-transpiler==0.11.0
qiskit-machine-learning==0.9.0
qiskit-nature==0.7.2
scikit-dimension
scikit-learn
scipy
seaborn
tensorflow
torch
tqdm
umap-learn
xgboost

# Documentation
sphinx
sphinx-autodoc-typehints 
sphinx_rtd_theme
myst_parser
better_apidoc
nbsphinx
myst_nb
furo
pydata-sphinx-theme
pandoc
sphinx-design
